design: 💄 textColorVariants를 Text.styles로 이동
Text 컴포넌트 파일 분리 및 index.ts 추가
