+++ src/components/atoms/Text/Text.types.ts
@@ -1,11 +1,9 @@
-import Link from "next/link";
 import { VariantProps } from "tailwind-variants";
 
-import { NonNullableProps } from "@customTypes/utilType";
+import type { PolymorphicPropsWithInnerRefType } from "@customTypes/polymorphicType";
+import type { NonNullableProps } from "@customTypes/utilType";
 
-import { textColorVariants } from "@utils/variants/testColorVariants";
-
-import { textVariants } from "./Text";
+import { textColorVariants, textVariants } from "@atoms/Text/Text.styles";
 
 /** 기본 Text 컴포넌트 타입 */
 export type TextDefault = "span";
@@ -15,7 +13,7 @@ export type TextDefault = "span";
  *
  * 필요 시 추가 후 사용
  */
-export type TextAlterAs = "div" | "a" | "p" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6" | typeof Link;
+export type TextAlterAs = "div" | "p" | "h1" | "h2" | "h3" | "h4" | "h5" | "h6";
 
 /**
  * Text 컴포넌트 기본 Props 타입
@@ -30,3 +28,8 @@ export type TextBasePropsType = NonNullableProps<
    */
   bold?: boolean;
 };
+
+export type TextProps<
+  T extends React.ElementType,
+  A extends React.ElementType = T
+> = PolymorphicPropsWithInnerRefType<T, TextBasePropsType, A>;
