+++ src/components/atoms/Text/index.ts
@@ -1,0 +1,11 @@
+import Text from "./Text";
+import { textColorVariants, textVariants } from "./Text.styles";
+import { TextAlterAs, TextDefault, TextProps } from "./Text.types";
+
+//export type
+export type { TextDefault, TextAlterAs, TextProps };
+
+//export styles
+export { textColorVariants, textVariants };
+
+export default Text;
